(c lib) Remove warnings at compile time

+++ eclipse/plugins/net.sf.orcc.backends/runtime/C/orcc/src/source.c
@@ -98,7 +98,7 @@ void source_init() {
 	loopsCount = nbLoops;
 }
 
-int source_open(char* fileName) {
+long source_open(char* fileName) {
 	stop = 0;
 	nb = 0;
 
@@ -117,7 +117,7 @@ int source_open(char* fileName) {
 	}
 	startTime = clock();
 	loopsCount = nbLoops;
-	return (int)file;
+	return (long)file;
 }
 
 unsigned int source_getNbLoop(void)
@@ -141,7 +141,7 @@ int source_sizeOfFile() {
 	return st.st_size; 
 }
 
-int source_sizeOfFileFd(int fdVal) {
+int source_sizeOfFileFd(long fdVal) {
 	FILE* fd = (FILE*) fdVal;
 	struct stat st;
 	fstat(fileno(fd), &st);
@@ -170,7 +170,7 @@ void source_rewind() {
 	}
 }
 
-void source_rewindFd(int fdVal) {
+void source_rewindFd(long fdVal) {
 	FILE* fd = (FILE*) fdVal;
 	if(fd != NULL) {
 		rewind(fd);
@@ -191,7 +191,7 @@ void source_close() {
 	}
 }
 
-void source_closeFd(int fdVal) {
+void source_closeFd(long fdVal) {
 	FILE* fd = (FILE*) fdVal;
 	if(fd != NULL) {
 		int n = fclose(fd);
@@ -235,7 +235,7 @@ void source_readNBytes(unsigned char *outTable, unsigned int nbTokenToRead){
 }
 
 
-void source_readNBytesFd(int fdVal, unsigned char *outTable, unsigned int nbTokenToRead){
+void source_readNBytesFd(long fdVal, unsigned char *outTable, unsigned int nbTokenToRead){
 	FILE* fd = (FILE*) fdVal;
 	int n = fread(outTable, 1, nbTokenToRead, fd);
